archiveInputs() now defaults to reading columns as character
This is to ensure that columns aren't overwritten with NA data when they are incorrectly read in. For example, if the third column of an inputs-archive CSV is text, but the col_types argument for read_csv() is "ccd", it will be read in as numeric, and as the text is not numeric, that column will return all NAs. When the file is re-written at the end of archiveInputs, the column will now be entirely NA.